Moved basic auth file to server middleware dir
