+++ packages/ripple-nuxt-tide/lib/server-middleware/basic-auth.js
